JFLAP and Hmmm autograders

+++ app/helpers/command.py
@@ -25,8 +25,13 @@ class Command(object):
             command = shlex.split(command)
         self.command = command
 
-    def run(self, timeout=None, **kwargs):
-        """ Run a command then return: (returncode, output, error). """
+    def run(self, timeout=None, compatibility=True, **kwargs):
+        """Run a command.
+
+        If compatibility mode is on (default), then return (timeout,
+        output, error). Else return (returncode, output, error,
+        timeout).
+        """
         def target(**kwargs):
             try:
                 self.process = subprocess.Popen(self.command, **kwargs)
@@ -47,6 +52,10 @@ class Command(object):
         timeout = False
         if thread.is_alive():
             timeout = True
-            self.process.terminate()
-            thread.join()
-        return timeout, self.output, self.error
+            if self.process:
+                self.process.terminate()
+                thread.join()
+        if compatibility:
+            return timeout, self.output, self.error
+        else:
+            return self.returncode, self.output, self.error, timeout
